Move ParaboloidMatVec out of paraboloid.py and into its own file in test_suite/components. Isn't used in first tutorial
