+++ openmdao/test_suite/components/paraboloid.py
@@ -38,32 +38,6 @@ class Paraboloid(ExplicitComponent):
         partials['f_xy', 'y'] = 2.0*y + 8.0 + x
 
 
-class ParaboloidMatVec(Paraboloid):
-    """ Use matrix-vector product instead."""
-
-    def compute_partials(self, inputs, outputs, partials):
-        """Analytical derivatives."""
-        pass
-
-    def compute_jacvec_product(self, inputs, outputs, dinputs, dresids, mode):
-        """Returns the product of the incoming vector with the Jacobian."""
-
-        x = inputs['x']
-        y = inputs['y']
-
-        if mode == 'fwd':
-            if 'x' in dinputs:
-                dresids['f_xy'] += (2.0*x - 6.0 + y)*dinputs['x']
-            if 'y' in dinputs:
-                dresids['f_xy'] += (2.0*y + 8.0 + x)*dinputs['y']
-
-        elif mode == 'rev':
-            if 'x' in dinputs:
-                dinputs['x'] += (2.0*x - 6.0 + y)*dresids['f_xy']
-            if 'y' in dinputs:
-                dinputs['y'] += (2.0*y + 8.0 + x)*dresids['f_xy']
-
-
 if __name__ == "__main__":
     from openmdao.core.problem import Problem
     from openmdao.core.group import Group
